+++ packages/medusa/.babelrc.js
@@ -7,16 +7,6 @@ if (process.env.NODE_ENV !== `test`) {
 }
 
 module.exports = {
-  plugins: [
-    "@babel/plugin-proposal-class-properties",
-    "@babel/plugin-transform-classes",
-    "@babel/plugin-transform-instanceof",
-  ],
-  presets: ["@babel/preset-env"],
-  env: {
-    test: {
-      plugins: ["@babel/plugin-transform-runtime"],
-    },
-  },
+  presets: [["babel-preset-medusa-package"]],
   ignore,
 }