+++ packages/medusa/src/api/routes/admin/orders/cancel-order.js
@@ -4,7 +4,11 @@ export default async (req, res) => {
   try {
     const orderService = req.scope.resolve("orderService")
     let order = await orderService.cancel(id)
-    order = await orderService.decorate(order, [], ["region"])
+    order = await orderService.decorate(
+      order,
+      [],
+      ["region", "customer", "swaps"]
+    )
     res.json({ order })
   } catch (error) {
     throw error