+++ packages/medusa/src/api/routes/admin/orders/receive-return.js
@@ -25,14 +25,18 @@ export default async (req, res) => {
     if (typeof value.refund !== "undefined" && value.refund < 0) {
       refundAmount = 0
     }
-    let order = await orderService.return(
+    let order = await orderService.receiveReturn(
       id,
       return_id,
       value.items,
       refundAmount,
       true
     )
-    order = await orderService.decorate(order, [], ["region"])
+    order = await orderService.decorate(
+      order,
+      [],
+      ["region", "customer", "swaps"]
+    )
 
     res.status(200).json({ order })
   } catch (err) {