+++ packages/medusa/src/api/routes/admin/orders/request-return.js
@@ -66,7 +66,11 @@ export default async (req, res) => {
       )
     }
 
-    order = await orderService.decorate(order, [], ["region"])
+    order = await orderService.decorate(
+      order,
+      [],
+      ["region", "customer", "swaps"]
+    )
 
     res.status(200).json({ order })
   } catch (err) {