+++ packages/medusa/src/api/routes/store/index.js
@@ -11,6 +11,8 @@ import orderRoutes from "./orders"
 import customerRoutes from "./customers"
 import shippingOptionRoutes from "./shipping-options"
 import regionRoutes from "./regions"
+import swapRoutes from "./swaps"
+import variantRoutes from "./variants"
 
 const route = Router()
 
@@ -35,6 +37,8 @@ export default (app, container, config) => {
   cartRoutes(route, container)
   shippingOptionRoutes(route)
   regionRoutes(route)
+  swapRoutes(route)
+  variantRoutes(route)
 
   return app
 }