+++ packages/medusa/src/services/customer.js
@@ -14,7 +14,7 @@ class CustomerService extends BaseService {
     CREATED: "customer.created",
   }
 
-  constructor({ customerModel, orderService, eventBusService }) {
+  constructor({ customerModel, eventBusService }) {
     super()
 
     /** @private @const {CustomerModel} */
@@ -22,9 +22,6 @@ class CustomerService extends BaseService {
 
     /** @private @const {EventBus} */
     this.eventBus_ = eventBusService
-
-    /** @private @const {EventBus} */
-    this.orderService_ = orderService
   }
 
   withSession(session) {
@@ -403,15 +400,6 @@ class CustomerService extends BaseService {
   async decorate(customer, fields = [], expandFields = []) {
     const requiredFields = ["_id", "metadata"]
     const decorated = _.pick(customer, fields.concat(requiredFields))
-
-    if (expandFields.includes("orders")) {
-      decorated.orders = await Promise.all(
-        customer.orders.map(async o => {
-          const order = await this.orderService_.retrieve(o)
-          return this.orderService_.decorate(order)
-        })
-      )
-    }
 
     const final = await this.runDecorators_(decorated)
     return final