+++ packages/medusa-core-utils/CHANGELOG.md
@@ -3,6 +3,120 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.11 (2020-11-24)
+
+
+
+## 1.0.10 (2020-09-09)
+
+
+### Bug Fixes
+
+* updates license ([db519fb](https://github.com/medusajs/medusa/commit/db519fbaa6f8ad02c19cbecba5d4f28ba1ee81aa))
+
+
+
+## 1.0.7 (2020-09-07)
+
+
+### Bug Fixes
+
+* **medusa-core-utils:** adds country utils ([00cbf84](https://github.com/medusajs/medusa/commit/00cbf8444d7b64ae8040db187748a63c64509686))
+
+
+
+## 1.0.1 (2020-09-05)
+
+
+
+## 1.0.1-beta.0 (2020-09-04)
+
+
+### Bug Fixes
+
+* **medusa:** product variant metadata ([#98](https://github.com/medusajs/medusa/issues/98)) ([520115a](https://github.com/medusajs/medusa/commit/520115a2cef7ee3f0259a682bd82e693c6327790))
+
+
+
+# 1.0.0 (2020-09-03)
+
+
+
+# 1.0.0-alpha.30 (2020-08-28)
+
+
+
+# 1.0.0-alpha.27 (2020-08-27)
+
+
+### Bug Fixes
+
+* **contentful-plugin:** Keep existing entry fields ([eb47896](https://github.com/medusajs/medusa/commit/eb478966684776bb2aa48e98789519644b05cd33))
+
+
+
+# 1.0.0-alpha.24 (2020-08-27)
+
+
+
+# 1.0.0-alpha.3 (2020-08-20)
+
+
+
+# 1.0.0-alpha.2 (2020-08-20)
+
+
+
+# 1.0.0-alpha.1 (2020-08-20)
+
+
+
+# 1.0.0-alpha.0 (2020-08-20)
+
+
+
+# 0.3.0 (2020-04-06)
+
+
+
+# 0.2.0 (2020-04-06)
+
+
+
+# 0.2.0-alpha.0 (2020-04-04)
+
+
+
+## 0.1.6-alpha.0 (2020-03-24)
+
+
+
+## 0.1.5-alpha.0 (2020-03-24)
+
+
+
+## 0.1.4-alpha.0 (2020-03-24)
+
+
+
+## 0.1.3-alpha.0 (2020-03-24)
+
+
+
+## 0.1.2-alpha.0 (2020-03-24)
+
+
+
+## 0.1.1-alpha.0 (2020-03-24)
+
+
+
+# 0.1.0-alpha.0 (2020-03-24)
+
+
+
+
+
 ## [1.0.10](https://github.com/medusajs/medusa/compare/v1.0.9...v1.0.10) (2020-09-09)