+++ packages/medusa-plugin-mailchimp/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.14](https://github.com/medusajs/medusa/compare/medusa-plugin-mailchimp@1.0.13...medusa-plugin-mailchimp@1.0.14) (2020-11-24)
+
+**Note:** Version bump only for package medusa-plugin-mailchimp
+
+
+
+
+
 ## [1.0.13](https://github.com/medusajs/medusa/compare/medusa-plugin-mailchimp@1.0.12...medusa-plugin-mailchimp@1.0.13) (2020-10-19)
 
 **Note:** Version bump only for package medusa-plugin-mailchimp